requirements.txt
anthropic==0.35.0
feedparser==6.0.10
google-api-python-client==1.12.5
trafilatura==1.8.0
youtube-transcript-api==0.6.2
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
